cast message template kind to int explicitly when saving it to wxConfig (fixes compilation with the latest wx svn)

+++ src/gui/wxTemplateDialog.cpp
@@ -959,7 +959,7 @@ bool wxAllTemplatesDialog::TransferDataFromWindow()
    wxConfigBase *config = wxConfigBase::Get();
    if ( config )
    {
-      config->Write(GetEditKindPath(), GetTemplateKind());
+      config->Write(GetEditKindPath(), (int)GetTemplateKind());
    }
 
    return true;
